change worker to SimpleWorker and change pytorch version

+++ encoder_worker/api.py
@@ -1,6 +1,6 @@
 from encoder import Encoder
 
 
-def encode_text(text: str) -> list[float]:
+def encode_text(text: str) -> list[int]:
     model = Encoder()
     return model(text=text)
